LPS-198099 Inline UserPersonalSitePermissions methods into CompanyModelListener
